vuepress class methods
